refactor: rename test file
